upstream_version(): allow specifying pypi name
This commit adds an optional pypi_name parameter to
upstream_version. This allows calling it multiple times for
specs that build packages from more than one pypi package and
thus cannot work with a single pypi_name context variable.

One example of this situation would be the monasca-ceilometer
package which is built from the monasca-ceilometer and
ceilometer pypi packages.